+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Logn
 All change logs will be documented here.<br/>
+   # [1.212] 2022-11-22
+   Fix bug: Field data type assigned incorrectly; warning in rmember/randstr. <br/>
    # [1.212] 2022-11-21
    Enhancement: Optimized the code of eval selection/sort expression and adding to result set (in matchFilter, group, setTree...). <br/>
    New feature: new query command: duplicate <num_expr>[ where filter_condition] -- Duplicate the fetched records. <br/>